+++ ui/src/com/kotcrab/vis/ui/widget/file/FileChooser.java
@@ -876,11 +876,16 @@ public class FileChooser extends VisWindow implements FileHistoryCallback {
 	protected void setStage (Stage stage) {
 		super.setStage(stage);
 
+		if (stage != null) {
+			refresh();
+		}
+
 		if (watchingFilesEnabled) {
-			if (stage != null)
+			if (stage != null) {
 				startFileWatcher();
-			else
+			} else {
 				stopFileWatcher();
+			}
 		}
 	}